backend/requirements.txt
FastAPI==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg[binary]>=3.1.0
python-telegram-bot==20.3
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
alembic==1.12.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx~=0.24.0
